requirements.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
billiard==4.2.4
boto3==1.35.70
botocore==1.35.99
celery==5.4.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.115.6
flake8==7.3.0
flake8-isort==7.0.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
isort==7.0.0
jmespath==1.0.1
kombu==5.6.2
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.2
pycodestyle==2.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
redis==5.2.0
rsa==4.9.1
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.30.6
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
